+++ src/lib.rs
@@ -2,7 +2,6 @@
 
 // Public modules
 mod common;
-mod error;
 
 #[cfg(feature = "async")]
 pub mod async_dht;
@@ -15,4 +14,3 @@ pub use bytes::Bytes;
 pub use dht::{Dht, Testnet};
 
 pub use ed25519_dalek::SigningKey;
-pub use error::{Error, Result};