+++ CHANGELOG.md
@@ -13,3 +13,12 @@ All notable changes to mainline dht will be documented in this file.
 - `Dht::shutdown()` is now idempotent, and returns `()`.
 - `AsyncDht::shutdown()` is now idempotent, and returns `()`.
 - `Rpc::drop` uses `tracing::debug!()` to log dropping the Rpc.
+- `Id::as_bytes()` instead of exposing internal `bytes` property.
+- Replace crate `Error` with more granular errors.
+- Replace Flume's `RecvError` with `expect()` message, since the sender should never be dropped to soon.
+- `DhtIsShutdown` error is a standalone error.
+- `InvalidIdSize` error is a standalone error.
+
+### Removed
+
+- Removed `mainline::error::Error` and `mainline::error::Result`